Replace ConfigParser by ConfigUpdater
